requirements.txt
Flask==0.12.2
numpy==1.13.1
Pillow==4.2.1
pygame==1.9.3
pyserial==3.3
scipy==0.19.1
